Refactored Search Suggestions
